refactoring and breaking things out
